+++ CHANGELOG.md
@@ -16,6 +16,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   [#76](https://github.com/ing-bank/cassandra-jdbc-wrapper/issues/76)).
 - Handle lists of User-Defined Types (UDT) properly when using `CassandraResultSet.getObject(String|int)` methods (see 
   issue [#76](https://github.com/ing-bank/cassandra-jdbc-wrapper/issues/76)).
+- Use an extended implementation of `UdtValue`, including a string representation of the value, for UDT values returned
+  by `CassandraResultSet.getObject(String|int)` methods (see issue 
+  [#76](https://github.com/ing-bank/cassandra-jdbc-wrapper/issues/76)).
 
 ## [4.14.0] - 2024-12-24
 ### Added